Fix/Test labelscan/base wine
- Add extra fields to base wine to reflect response
- Fix double/ int issue with rating summary

+++ Delectable/app/src/main/java/com/delectable/mobile/api/models/RatingsSummaryHash.java
@@ -14,11 +14,11 @@ public class RatingsSummaryHash {
         return pro != null ? pro.count : 0;
     }
 
-    public int getAllAvg() {
+    public double getAllAvg() {
         return all != null ? all.avg : 0;
     }
 
-    public int getProAvg() {
+    public double getProAvg() {
         return pro != null ? pro.avg : 0;
     }
 
@@ -50,7 +50,7 @@ public class RatingsSummaryHash {
 
         Integer count;
 
-        Integer avg;
+        Double avg;
 
         @Override
         public String toString() {
